Reorganize PWA components and improve mobile installation experience
- Restructure PWA components into organized pwa/ folder following app patterns
- Add index.ts for clean component exports
- Enhance mobile PWA installation with intelligent device detection
- Implement native Android install prompts when available
- Add device-specific installation instructions for iOS and Android
- Fix banner positioning and prevent horizontal scrolling on mobile
- Improve install banner state management to prevent flickering
- Update navigation layout CSS for better responsive behavior

Components moved:
- src/lib/components/ui/pwa-status.svelte -> src/lib/components/ui/pwa/pwa-status.svelte
- src/lib/components/ui/pull-to-refresh.svelte -> src/lib/components/ui/pwa/pull-to-refresh.svelte
- src/lib/components/ui/offline-page.svelte -> src/lib/components/ui/pwa/offline-page.svelte
